# packages/backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sistem POS API",
    "version": "1.0.0",
    "description": "Dokumentasi REST API Sistem POS real-time multi-tenant.",
    "contact": {
      "name": "Sistem POS Team",
      "email": "support@sistempos.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 120
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "totalPages": {
            "type": "integer",
            "example": 12
          },
          "limit": {
            "type": "integer",
            "example": 10
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Terjadi kesalahan"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/produk": {
      "get": {
        "tags": [
          "Produk"
        ],
        "summary": "Cari produk",
        "description": "Mencari produk dengan filter dan pagination",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string",
              "default": "1"
            },
            "description": "Nomor halaman"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string",
              "default": "10"
            },
            "description": "Jumlah item per halaman"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Kata kunci pencarian"
          },
          {
            "in": "query",
            "name": "kategori_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter berdasarkan kategori"
          },
          {
            "in": "query",
            "name": "brand_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter berdasarkan brand"
          },
          {
            "in": "query",
            "name": "is_aktif",
            "schema": {
              "type": "string",
              "enum": [
                "0",
                "1"
              ]
            },
            "description": "Filter status aktif"
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar produk berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Produk"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Produk"
        ],
        "summary": "Buat produk baru",
        "description": "Membuat produk baru (memerlukan akses toko)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "kategori_id",
                  "brand_id",
                  "supplier_id",
                  "kode",
                  "nama"
                ],
                "properties": {
                  "kategori_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID kategori produk"
                  },
                  "brand_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID brand produk"
                  },
                  "supplier_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID supplier produk"
                  },
                  "kode": {
                    "type": "string",
                    "maxLength": 50,
                    "description": "Kode produk unik"
                  },
                  "nama": {
                    "type": "string",
                    "maxLength": 255,
                    "description": "Nama produk"
                  },
                  "harga_beli": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Harga beli produk"
                  },
                  "harga_jual": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Harga jual produk"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produk berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Produk"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Data tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/produk/{id}": {
      "get": {
        "tags": [
          "Produk"
        ],
        "summary": "Ambil produk berdasarkan ID",
        "description": "Mengambil detail produk berdasarkan ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID produk"
          }
        ],
        "responses": {
          "200": {
            "description": "Detail produk berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Produk"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Produk"
        ],
        "summary": "Update produk",
        "description": "Mengupdate data produk berdasarkan ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID produk"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nama": {
                    "type": "string",
                    "maxLength": 255,
                    "description": "Nama produk"
                  },
                  "harga_beli": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Harga beli produk"
                  },
                  "harga_jual": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Harga jual produk"
                  },
                  "is_aktif": {
                    "type": "integer",
                    "enum": [
                      0,
                      1
                    ],
                    "description": "Status aktif produk"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produk berhasil diupdate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Produk"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Data tidak valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Produk"
        ],
        "summary": "Hapus produk",
        "description": "Menghapus produk berdasarkan ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID produk"
          }
        ],
        "responses": {
          "200": {
            "description": "Produk berhasil dihapus",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Product deleted successfully"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Gagal menghapus produk",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}